LayoutTests/fast/mediastream/RTCIceCandidate.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <script>
            description("Tests RTCIceCandidate.");

            var initializer = {candidate:"foo", sdpMid:"bar", sdpMLineIndex:6};
            var candidate;
            shouldNotThrow('candidate = new RTCIceCandidate(initializer);');
            shouldBe('candidate.candidate', '"foo"');
            shouldBe('candidate.sdpMid', '"bar"');
            shouldBe('candidate.sdpMLineIndex', '6');
            debug("");

            debug("Attributes are readonly");
            evalAndLog('candidate.candidate = "foo-updated"');
            shouldBe('candidate.candidate', '"foo"');
            evalAndLog('candidate.sdpMid = "bar-updated"');
            shouldBe('candidate.sdpMid', '"bar"');
            evalAndLog('candidate.sdpMLineIndex = 7');
            shouldBe('candidate.sdpMLineIndex', '6');
            debug("");

            debug("Dictionary argument is mandatory.");
            shouldThrow('new RTCIceCandidate();');

            debug("Dictionary member 'candidate' is required.");
            shouldThrow('new RTCIceCandidate({});');
            shouldThrow('new RTCIceCandidate({sdpMid:"bar", sdpMLineIndex:6});');
            shouldThrow('new RTCIceCandidate({sdpMid:"bar"});');
            shouldThrow('new RTCIceCandidate({sdpMLineIndex:6});');
            debug("");

            debug("One of the 'sdpMid' or 'sdpMLineIndex' members must be present (and not null or undefined).");
            shouldThrow('new RTCIceCandidate({candidate:"foo"});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMid: null});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMid: undefined});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMLineIndex: null});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMLineIndex: undefined});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMid: null, sdpMLineIndex: null});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMid: undefined, sdpMLineIndex: undefined});');
            debug("");

            debug("Test that sdpMid or sdpMLineIndex explicitly set to null or undefined, results in a null value.");
            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMid: null, sdpMLineIndex: 1});');
            shouldBeNull("candidate.sdpMid");
            shouldBe("candidate.sdpMLineIndex", "1");

            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMid: undefined, sdpMLineIndex: 1});');
            shouldBeNull("candidate.sdpMid");
            shouldBe("candidate.sdpMLineIndex", "1");

            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMid: "foo", sdpMLineIndex: null});');
            shouldBe("candidate.sdpMid", "'foo'");
            shouldBeNull("candidate.sdpMLineIndex");

            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMid: "foo", sdpMLineIndex: undefined});');
            shouldBe("candidate.sdpMid", "'foo'");
            shouldBeNull("candidate.sdpMLineIndex");
            debug("");

            debug("When one of the 'sdpMid' or 'sdpMLineIndex' members is set, the other must be null.");
            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMid:"bar"});');
            shouldBe('candidate.sdpMid', '"bar"');
            shouldBe('candidate.sdpMLineIndex', 'null');
            shouldNotThrow('candidate = new RTCIceCandidate({candidate:"foo", sdpMLineIndex:6});');
            shouldBe('candidate.sdpMid', 'null');
            shouldBe('candidate.sdpMLineIndex', '6');
            debug("");

            debug("Some bad input");
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMLineIndex:"a"});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMLineIndex:-1});');
            shouldThrow('new RTCIceCandidate({candidate:"foo", sdpMLineIndex:65536});');
            shouldThrow('new RTCIceCandidate(5);');
            shouldThrow('new RTCIceCandidate("foobar");');
            shouldThrow('new RTCIceCandidate(candidate:"");');

            window.successfullyParsed = true;
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
